set_device and float32 edits for snmg pagerank

+++ python/cugraph/mg_pagerank/mg_pagerank_wrapper.pyx
@@ -24,7 +24,7 @@ def mg_pagerank(src_ptrs_info, dest_ptrs_info):
 
     data = rmm.device_array_from_ptr(<uintptr_t> pr_ptr.data,
                                             nelem=pr_ptr.size,
-                                            dtype=np.int32)
+                                            dtype=np.float32)
     x = cudf.DataFrame()
     x['vertex'] = np.ones(pr_ptr.size,dtype=int)
     x['pagerank'] = cudf.Series(data)
